tests(brand_new_tests): Architecture
